+++ mbgdml/predictors/gdml_predict.py
@@ -24,6 +24,8 @@
 import logging
 import numpy as np
 
+from ..stress import virial_atom_loop
+
 log = logging.getLogger(__name__)
 
 # This calculation is too fast to be a ray task.
@@ -198,8 +200,12 @@ def predict_gdml(Z, R, entity_ids, entity_combs, model, periodic_cell, **kwargs)
     F = np.zeros(R.shape)
 
     alchemy_scalers = kwargs.get("alchemy_scalers", None)
+    compute_virial = kwargs.get("compute_virial", False)
 
     periodic = bool(periodic_cell)
+    compute_virial = bool(periodic and compute_virial)
+    if compute_virial:
+        virial = np.zeros((3, 3), dtype=np.float64)
 
     # Getting all contributions for each molecule combination (comb).
     for entity_id_comb in entity_combs:
@@ -211,15 +217,18 @@ def predict_gdml(Z, R, entity_ids, entity_combs, model, periodic_cell, **kwargs)
             r_slice.extend(np.where(entity_ids == entity_id)[0])
 
         z = Z[r_slice]
-        r = R[r_slice]
+        r_orig = R[r_slice]
+        # Note: We store the original coordinates in case the virial is requested
 
         # If we are using a periodic cell we convert r_comp into coordinates
         # we can use in many-body expansions.
         if periodic:
-            r = periodic_cell.r_mic(r)
+            r = periodic_cell.r_mic(r_orig)
             if r is None:
                 # Any atomic pairwise distance was larger than cutoff.
                 continue
+        else:
+            r = r_orig
 
         # TODO: Check if we can avoid prediction if we have an alchemical factor of
         # zero?
@@ -259,6 +268,12 @@ def predict_gdml(Z, R, entity_ids, entity_combs, model, periodic_cell, **kwargs)
         # Adds contributions to total energy and forces.
         E += e
         F[r_slice] += f
+
+        if compute_virial:
+            virial += virial_atom_loop(r_orig, f)
+
+    if compute_virial:
+        return E, F, virial
 
     return E, F